+++ tests/recipes/regression/v1/test_regression_pipeline.py
@@ -4,11 +4,6 @@ import pytest
 from mlflow.exceptions import MlflowException
 from mlflow.recipes.regression.v1.recipe import RegressionRecipe
 
-# pylint: disable=unused-import
-from tests.recipes.helper_functions import (
-    enter_recipe_example_directory,
-)  # pylint: enable=unused-import
-
 
 @pytest.fixture
 def create_recipe(enter_recipe_example_directory):
