+++ tests/recipes/test_ingest_step.py
@@ -16,13 +16,6 @@ from mlflow.store.artifact.s3_artifact_repo import S3ArtifactRepository
 from mlflow.utils.file_utils import read_yaml
 from mlflow.recipes.utils import _RECIPE_CONFIG_FILE_NAME
 
-# pylint: disable=unused-import
-from tests.recipes.helper_functions import (
-    tmp_recipe_root_path,
-    enter_recipe_example_directory,
-    enter_test_recipe_directory,
-)
-
 
 @pytest.fixture
 def pandas_df():
@@ -870,8 +863,6 @@ def test_ingests_spark_sql_datetime_successfully(spark_session, tmp_path):
     from pyspark.sql.functions import (
         col,
         rand,
-        lit,
-        date_sub,
         unix_timestamp,
         to_timestamp,
         current_date,
